リポジトリを削除する際に、確認ダイアログが表示されるように修正

+++ app/javascript/repository-delete-button.js
@@ -1,15 +1,18 @@
-document.addEventListener('DOMContentLoaded', () => {
+document.addEventListener('turbo:load', () => {
   const repositoryDeleteButton = document.querySelector('#repository-delete-button')
   repositoryDeleteButton.addEventListener('click', () => {
     const selectedRepository = document.querySelector('option:checked')
-    fetch(`/repositories/${selectedRepository.getAttribute('value')}`, {
-      method: 'DELETE',
-      headers: headers(),
-      credentials: 'same-origin',
-      redirect: 'follow'
-    }).then(res => {
-      location.href = res.url
-    })
+    const res = confirm('本当に実行してもよろしいですか？')
+    if(res) {
+      fetch(`/repositories/${selectedRepository.getAttribute('value')}`, {
+        method: 'DELETE',
+        headers: headers(),
+        credentials: 'same-origin',
+        redirect: 'follow'
+      }).then(res => {
+        location.href = res.url
+      })
+    }
   })
 })
